+++ src/app.d.ts
@@ -6,8 +6,9 @@ import type { SvelteComponent } from 'svelte'
 declare global {
 	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	type Componenet = typeof SvelteComponent<any, any, any>
-	// Yes its deprecated however .currentDirection does not get set to stopped in Firefox when stopped while .stopped is still supported everywhere.
+
 	interface RTCRtpTransceiver {
+		// yes its deprecated however `currentDirection` does not get set to stopped in Firefox when stopped while `stopped` is still supported everywhere.
 		readonly stopped: boolean
 	}