+++ src/lib/components/secondary-panel/tab-contacts.svelte
@@ -3,5 +3,5 @@
 </script>
 
 <SubGroup>
-	<div class="mx-1 h-40">TODO</div>
+	<div class="mx-1 h-40">Soon™</div>
 </SubGroup>