+++ src/lib/components/secondary-panel/tab-soundboard.svelte
@@ -1,7 +1,0 @@
-<script lang="ts">
-	import SubGroup from '../core/sub-group.svelte'
-</script>
-
-<SubGroup>
-	<div class="mx-1 h-40">TODO</div>
-</SubGroup>